+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - improved waiting for server: `OnlineOptions.connectionTimeout`
   is now also used when connecting to the server fails
 - added `Address.extension`
+- added `Administration.shutdown` and `DomainAdministration.shutdown*`
 
 ## 1.1.0 [The Wintermute Edition](https://en.wikipedia.org/wiki/Neuromancer) (2016-04-25)